added abstraction
