+++ solution/backend/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/client.ts","./src/index.ts"],"version":"5.8.3"}
+{"root":["./src/client.ts","./src/index.ts","./src/services/generate-question.ts"],"version":"5.8.3"}
